tests/requirements-test.txt
pytest
pytest-flask
coverage
flake8
pylint
isort
black
